requirements.txt
# Function dependencies
google-api-core
google-cloud-securitycenter
google-cloud-storage